feat: Decreasing code duplication
