+++ src/modules/75-ci/components/PipelineSteps/RunStep/RunStepBase.tsx
@@ -6,7 +6,6 @@
  */
 
 import React from 'react'
-import { useParams } from 'react-router-dom'
 import cx from 'classnames'
 import {
   Text,
@@ -16,21 +15,17 @@ import {
   FormikForm,
   Accordion,
   Color,
-  Container,
-  Layout
+  Container
 } from '@wings-software/uicore'
 import type { FormikProps } from 'formik'
 import get from 'lodash/get'
 import type { K8sDirectInfraYaml } from 'services/ci'
-import { Connectors } from '@connectors/constants'
-import { FormMultiTypeConnectorField } from '@connectors/components/ConnectorReferenceField/FormMultiTypeConnectorField'
 import { StepFormikFowardRef, setFormikRef } from '@pipeline/components/AbstractSteps/Step'
 import MultiTypeFieldSelector from '@common/components/MultiTypeFieldSelector/MultiTypeFieldSelector'
 import { usePipelineContext } from '@pipeline/components/PipelineStudio/PipelineContext/PipelineContext'
 import { useStrings } from 'framework/strings'
 import { ConfigureOptions } from '@common/components/ConfigureOptions/ConfigureOptions'
 import { ShellScriptMonacoField } from '@common/components/ShellScriptMonaco/ShellScriptMonaco'
-import { MultiTypeTextField } from '@common/components/MultiTypeText/MultiTypeText'
 import StepCommonFields, {
   GetImagePullPolicyOptions,
   GetShellOptions
@@ -41,16 +36,16 @@ import {
   getFormValuesInCorrectFormat
 } from '@pipeline/components/PipelineSteps/Steps/StepsTransformValuesUtils'
 import { validate } from '@pipeline/components/PipelineSteps/Steps/StepsValidateUtils'
-import { useGitScope } from '@pipeline/utils/CIUtils'
 import type { RunStepProps, RunStepData, RunStepDataUI } from './RunStep'
 import { transformValuesFieldsConfig, getEditViewValidateFieldsConfig } from './RunStepFunctionConfigs'
-import { CIStepOptionalConfig, getOptionalSubLabel } from '../CIStep/CIStepOptionalConfig'
+import { CIStepOptionalConfig } from '../CIStep/CIStepOptionalConfig'
 import {
   AllMultiTypeInputTypesForStep,
   useGetPropagatedStageById,
   validateConnectorRefAndImageDepdendency
 } from '../CIStep/StepUtils'
 import { CIStep } from '../CIStep/CIStep'
+import { AWSVMBuildInfraCommon } from '../CIStep/AWSVMBuildInfraCommon'
 import css from '@pipeline/components/PipelineSteps/Steps/Steps.module.scss'
 
 export const RunStepBase = (
@@ -65,87 +60,10 @@ export const RunStepBase = (
 
   const { getString } = useStrings()
   const { expressions } = useVariablesExpression()
-  const gitScope = useGitScope()
-  const { accountId, projectIdentifier, orgIdentifier } = useParams<{
-    projectIdentifier: string
-    orgIdentifier: string
-    accountId: string
-  }>()
 
   const currentStage = useGetPropagatedStageById(selectedStageId || '')
 
   const buildInfrastructureType = get(currentStage, 'stage.spec.infrastructure.type') as K8sDirectInfraYaml['type']
-
-  const renderConnectorRefAndImage = React.useCallback(
-    (showOptionalSublabel: boolean) => (
-      <>
-        <Container className={css.bottomMargin3}>
-          <FormMultiTypeConnectorField
-            label={
-              <Layout.Horizontal flex={{ justifyContent: 'flex-start', alignItems: 'baseline' }}>
-                <Text
-                  className={css.inpLabel}
-                  color={Color.GREY_600}
-                  font={{ size: 'small', weight: 'semi-bold' }}
-                  style={{ display: 'flex', alignItems: 'center' }}
-                >
-                  {getString('pipelineSteps.connectorLabel')}
-                </Text>
-                &nbsp;
-                {showOptionalSublabel ? getOptionalSubLabel('', getString) : null}
-              </Layout.Horizontal>
-            }
-            type={[Connectors.GCP, Connectors.AWS, Connectors.DOCKER]}
-            width={385}
-            name={`spec.connectorRef`}
-            placeholder={getString('select')}
-            accountIdentifier={accountId}
-            projectIdentifier={projectIdentifier}
-            orgIdentifier={orgIdentifier}
-            multiTypeProps={{
-              expressions,
-              allowableTypes: AllMultiTypeInputTypesForStep,
-              disabled: readonly
-            }}
-            gitScope={gitScope}
-            setRefValue
-          />
-        </Container>
-        <Container className={cx(css.formGroup, css.lg, css.bottomMargin5)}>
-          <MultiTypeTextField
-            name={`spec.image`}
-            label={
-              <Layout.Horizontal flex={{ justifyContent: 'flex-start', alignItems: 'baseline' }}>
-                <Text
-                  className={css.inpLabel}
-                  color={Color.GREY_600}
-                  font={{ size: 'small', weight: 'semi-bold' }}
-                  tooltipProps={
-                    showOptionalSublabel
-                      ? {}
-                      : {
-                          dataTooltipId: 'image'
-                        }
-                  }
-                  placeholder={getString('imagePlaceholder')}
-                >
-                  {getString('imageLabel')}
-                </Text>
-                &nbsp;
-                {showOptionalSublabel ? getOptionalSubLabel('image', getString) : null}
-              </Layout.Horizontal>
-            }
-            multiTextInputProps={{
-              multiTextInputProps: {
-                allowableTypes: AllMultiTypeInputTypesForStep
-              }
-            }}
-          />
-        </Container>
-      </>
-    ),
-    []
-  )
 
   return (
     <Formik
@@ -206,7 +124,7 @@ export const RunStepBase = (
                 description: {}
               }}
             />
-            {buildInfrastructureType !== 'VM' ? <>{renderConnectorRefAndImage(false)}</> : null}
+            {buildInfrastructureType !== 'VM' ? <AWSVMBuildInfraCommon showOptionalSublabel={false} /> : null}
             <div className={cx(css.fieldsGroup, css.withoutSpacing, css.topPadding3, css.bottomPadding3)}>
               <MultiTypeFieldSelector
                 name="spec.command"
@@ -266,7 +184,7 @@ export const RunStepBase = (
                 summary={getString('common.optionalConfig')}
                 details={
                   <Container margin={{ top: 'medium' }}>
-                    {buildInfrastructureType === 'VM' ? renderConnectorRefAndImage(true) : null}
+                    {buildInfrastructureType === 'VM' ? <AWSVMBuildInfraCommon showOptionalSublabel={true} /> : null}
                     <CIStepOptionalConfig
                       stepViewType={stepViewType}
                       readonly={readonly}
